Moved LEO functions out of KB and into LEO.java
