+++ src/java/com/articulate/sigma/Diagnostics.java
@@ -1322,11 +1322,13 @@ public class Diagnostics {
                 //System.out.println(" processedQueries = " + processedQueries);
 
                 System.out.println("INFO in Diagnostics.kbConsistencyCheck(): size = " + processedQueries.size());
+                EProver eprover = new EProver();
                 for (Formula f : processedQueries) {
                     System.out.println("INFO in Diagnostics.kbConsistencyCheck(): formula = " + f.getFormula());
                     processedQuery = f.makeQuantifiersExplicit(false);
                     System.out.println("INFO in Diagnostics.kbConsistencyCheck(): processedQuery = " + processedQuery);
-                    proof = EProver.askEProver(kb, processedQuery, "tptp", timeout, maxAnswers).toString() + " ";
+                    eprover.askEProver(kb, processedQuery, "tptp", timeout, maxAnswers);
+                    proof = eprover.toString() + " ";
                     a = new StringBuilder();
                     a.append(reportAnswer(kb,proof,query,processedQuery,"Redundancy"));
                     //  if (answer.length() != 0) return answer;
@@ -1334,7 +1336,8 @@ public class Diagnostics {
 
                     negatedQuery = new StringBuilder();
                     negatedQuery.append(Formula.LP).append(Formula.NOT).append(Formula.SPACE).append(processedQuery).append(Formula.RP);
-                    proof = EProver.askEProver(kb, negatedQuery.toString(), "tptp", timeout, maxAnswers).toString() + " ";
+                    eprover.askEProver(kb, negatedQuery.toString(), "tptp", timeout, maxAnswers);
+                    proof = eprover.toString() + " ";
                     a.append(reportAnswer(kb,proof,query,negatedQuery.toString(),"Inconsistency"));
                     if (a.length() != 0) {
                         answer.append(a);
